move RobotSkills to Skills namespace
