+++ Assets/Dev/Arby/core_actions/Scripts/CoreActions/Controllers/RobotActionController.cs
@@ -1,3 +1,4 @@
+using RobotCoreAction.Skills;
 using UnityEngine;
 
 namespace RobotCoreAction
