+++ Assets/Dev/Arby/core_actions/Scripts/CoreActions/Skills/RobotSkills.cs
@@ -1,0 +1,20 @@
+using RobotCoreAction.Controllers;
+
+namespace RobotCoreAction
+{
+    namespace Skills
+    {
+        public interface ISkill
+        {
+            ERobotSkillType SkillType { get; }
+            float Duration { get; }
+            float Cooldown { get; }
+            void Execute(RobotActionController controller, RobotStats stats, RobotPhysicController physic);
+        }
+        public enum ERobotSkillType
+        {
+            Stone,
+            Boost
+        }
+    }
+}
